rename runner directory to host_manager
